Fix warning from signed/unsigned mismatch
Seen (and breaking local build) with VC++2022:

qrhid3d12_p_p.h(206): warning C4018: '<': signed/unsigned mismatch

QD3D12ObjectHandle::index is a quint32, whereas data is a QVector,
with count() returning a signed qsizetype.

Pick-to: 6.5

+++ src/gui/rhi/qrhid3d12_p_p.h
@@ -203,7 +203,7 @@ struct QD3D12ObjectPool
     bool isValid(const QD3D12ObjectHandle &handle) const
     {
         return handle.index > 0
-                && handle.index < data.count()
+                && handle.index < quint32(data.count())
                 && handle.generation > 0
                 && handle.generation == data[handle.index].generation
                 && data[handle.index].object.has_value();
